@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace MarcelStrahl\PriceCalculator;
 
 use MarcelStrahl\PriceCalculator\Helpers\Entity\Price;
@@ -55,7 +57,7 @@ class PriceCalculator implements PriceCalculatorInterface
             return $price;
         }
 
-        $price->setPrice($price->getPrice() / $amount);
+        $price->setPrice((int) ($price->getPrice() / $amount));
 
         return $price;
     }
